Add ticket logging functionality
Fully implemented ticket logging in the TM. The TM now keeps a log of actual ticket start and end times. It saves that info along with the ticket details every 5 minutes while it is running. On shutdown, it saves the log to the same file. This ensures there's only one log file per TM run.

TM also holds a log service which will be used by the SG for schedule visualization.