Update conversion_utils.py
Old task dict to ticket list function and its reverse function depended heavily on the order of attributes to ensure no errors occurred in the calling nodes.

Implementing the logging functionality increased the number of places where this order needed to be different, so I removed the order dependency. Now, we only add the info if it's in the ticket or message. Initial testing suggests this works as expected. Will monitor.