Fixed ticket_log bugs from a 09/26 commit
In a previous commit, I changed "duration" in ticket_log to "estimated_duration" but this causes downstream errors when the TM sends logs to the SG. Changed it back to "duration"

Also had changed start and end to human-readable times, but this causes errors as well when sending logs. Instead, now they're only converted when saving the logs (check the ticket_copy portion on line 815 onwards).